Add status mapping constants and implement status management features: create StatusForm and StatusFormEdit components for handling status updates, integrate them into the dashboard, and refactor related components for improved request management.
