+++ src/app/dashboard/home/hooks/useAllSolicitudes.query.ts
@@ -1,9 +1,6 @@
 "use client";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import {
-  getSolicitudes,
-  updateSolicitud,
-} from "../services/solicitudes.service";
+import { useQuery } from "@tanstack/react-query";
+import { getSolicitudes } from "../services/solicitudes.service";
 import { decodeUserToken } from "@app/(global)/utils/decode-user-token.util";
 
 export const useAllSolicitudes = () => {
@@ -13,22 +10,6 @@ export const useAllSolicitudes = () => {
       const userData = decodeUserToken();
       const solicitudesData = await getSolicitudes(userData?.user_id ?? "");
       return solicitudesData.solicitudes;
-    },
-  });
-};
-
-export const useUpdateSolicitud = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationKey: ["updateSolicitud"],
-    mutationFn: (updateData: {
-      solicitudId: string;
-      status: number;
-      observation?: string;
-      fileUrl?: string;
-    }) => updateSolicitud(updateData),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["allSolicitudes"] });
     },
   });
 };
