+++ src/app/dashboard/home/utils/solicitudes-en-proceso.util.tsx
@@ -1,7 +1,7 @@
 import { formatDate } from "./format-date.util";
 import { solicitudesFiltradas } from "./solicitudes-filtradas.util";
 import { STATUS_CLASSES } from "../constants/status-classes.constant";
-import { STATUS_MAPPING } from "../constants/status-mapping.constant";
+import { STATUS_MAPPING } from "../../../(global)/constants/status-mapping.constant";
 import { TIPO_MAPPING } from "../constants/tipo-mapping.constant";
 
 export const solicitudesEnProceso = (allSolicitudes: any[], formData: any) => {
